refactor(auto_source): encapsulate classifiers behind LinkHeuristics assessment interface
